Create method to calculate dot product of 2 vectors
1.  Create CalcDotProduct() static method

2.  Create unit test method for testing the CalcDotProduct() static method.

3.  Change result of CalcLength unit tests to use ints.

+++ KDPhysicsTest/Vect2Tests.cs
@@ -31,7 +31,7 @@ namespace KDPhysicsTest
             var expected = 22;
 
             //Act
-            var actual = vector.Length;
+            var actual = (int)vector.Length;
 
             //Assert
             Assert.AreEqual(expected, actual);
@@ -45,7 +45,7 @@ namespace KDPhysicsTest
             var expected = 22;
 
             //Act
-            var actual = Vect2.CalcLength(vector);
+            var actual = (int)Vect2.CalcLength(vector);
 
             //Assert
             Assert.AreEqual(expected, actual);
